+++ install_miniforge.bash
@@ -686,6 +686,8 @@ fi
 $PIP_INSTALL redis
 $PIP_INSTALL Flask
 $PIP_INSTALL goes2go
+$PIP_INSTALL nco
+$PIP_INSTALL cdo
 
 # some packages require a Fortran compiler. This sometimes isn't available
 # on macs (though usually is)