+++ install_miniforge.bash
@@ -605,10 +605,6 @@ $PACKAGE_INSTALL uxarray
 
 $PACKAGE_INSTALL contextily
 
-# We seem to need to require sqlite 3.48.0 *exactly* for ipython3
-# NOTE: This might need to be revisited in the next version
-#$PACKAGE_INSTALL sqlite"==3.48.0"
-
 # Only install pythran on linux. On mac it brings in an old clang
 if [[ $MINIFORGE_ARCH == Linux ]]
 then