+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Update example Miniforge version to 25.3.0-3
+- Update example Python version to 3.13
+- Re-enable `basemap` installation by default
+  - This is now possible as `basemap` has been updated to support `numpy` v2
+
 ### Added
 
 - Explicit Pip Packages