+++ CHANGELOG.md
@@ -11,14 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Update example Miniforge version to 25.3.0-3
+- Update example Miniforge version to 25.3.1-3
 - Update example Python version to 3.13
 - Re-enable `basemap` installation by default
   - This is now possible as `basemap` has been updated to support `numpy` v2
 - Rename `always_circular_stereo.py` to `cartopy_example.py`
+- Enable TensorFlow installation for Python 3.13
+  - This is now possible as TensorFlow has been updated to support Python 3.13
 
 ### Added
 
+- Explicit Conda Packages
+  - contextily
 - Explicit Pip Packages
   - nco
   - cdo